Remove updatePlayButton.ts
